compile without database support
